% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sidebox.R
\name{consort_plot_sidebox}
\alias{consort_plot_sidebox}
\title{Find Side Box from \code{orders} Argument}
\usage{
consort_plot_sidebox(data, orders, sidebox_pattern = "^sidebox_", ...)
}
\arguments{
\item{data}{\link[base]{data.frame}}

\item{orders}{\link[base]{character} \link[base]{vector}}

\item{sidebox_pattern}{\link[base]{regex}}

\item{...}{additional parameters of \link[consort]{consort_plot}}
}
\value{
Function \link{consort_plot_sidebox} returns a \CRANpkg{consort} object.
}
\description{
Find Side Box from \code{orders} Argument
}
\details{
Function \link{consort_plot_sidebox} finds the argument \code{side_box} of
function \link[consort]{consort_plot}, from the argument of \code{orders}.
}
